==> 1_Web Scraping.ipynb <==
# Lab | Web Scraping Single Page

from bs4 import BeautifulSoup

# access information on any webpage...

from bs4 import BeautifulSoup
import requests
import pandas as pd

pd.set_option('display.max_columns', 165)
pd.set_option('display.max_rows', 165)

# 2. find url and store it in a variable
# top 100 songs - 2015
url = "https://playback.fm/charts/top-100-songs/2015"

# 3. download html with a get request
response = requests.get(url)
# response.status_code # 200 status code means OK!

response.status_code # 200 status code means OK!

# response.content

# 4.1. parse html (create the 'soup')
soup = BeautifulSoup(response.content, "html.parser")

# 4.2. check that the html code looks like it should
# soup

# 5. retrieve/extract the desired info (here, you'll paste the "Selector" you copied before to get the element that belongs to the top movie)
# soup.select("td:nth-child(2)")



# print(soup.prettify())

soup.select("td:nth-child(2) > a")[0]

soup.select("td:nth-child(2) > a")[99].get_text() # artist

soup.select("td:nth-child(2)")[0]

soup.select("td.mobile-hide > a > span.song")[0].get_text() #song

#initialize empty lists
artist = []
song = []

# define the number of iterations of our for loop
# by checking how many elements are in the retrieved result set
# (this is equivalent but more robust than just explicitly defining 250 iterations)
num_iter = len(soup.select("td:nth-child(2) > a"))

tClist = soup.select("td:nth-child(2) > a")
spanlist = soup.select("td.mobile-hide > a > span.song")
# iterate through the result set and retrive all the data
for i in range(num_iter):
    artist.append(tClist[i].get_text())
    song.append(spanlist[i].get_text())

print(num_iter)
print(artist)
print(song)



# each list becomes a column
songs = pd.DataFrame({"artist":artist,
                       "song":song, 
                      })

songs

songs['artist']=songs['artist'].str.replace('\n', ' ')
             

songs.head()

len(songs)

import random


secure_random = random.SystemRandom()
print(secure_random.choice(song))

# print(songs['song'].sample())

favorite_song = input('Please enter your song name: ')


print(favorite_song)

# Recommend song

if len(songs[(songs['song'].isin([favorite_song]))])>0:
    print("Our song recommendation is: ", secure_random.choice(songs['song'])) 
else:
    print("We are sorry that we do not have a song recommendation for you now.")

songs_1=songs

# Lab | Web Scraping Multiple Page

## Top-100-Songs

# 2. find url and store it in a variable
# Bilboard hot 100
url = "https://www.popvortex.com/music/charts/top-100-songs.php"

# 3. download html with a get request
response = requests.get(url)
# response.status_code # 200 status code means OK!

response.status_code # 200 status code means OK!

# response.content

# 4.1. parse html (create the 'soup')
soup = BeautifulSoup(response.content, "html.parser")

# print(soup.prettify())

soup.select("div.chart-content.col-xs-12.col-sm-8 > p > em")[0].get_text() 

# artist

soup.select("div.chart-content.col-xs-12.col-sm-8 > p > cite")[0].get_text() 

# song

#chart-position-1 > div.chart-content.col-xs-12.col-sm-8 > p > cite # song

#chart-position-1 > div.chart-content.col-xs-12.col-sm-8 > p > em # artist

#initialize empty lists
artist = []
song = []

# define the number of iterations of our for loop
# by checking how many elements are in the retrieved result set
# (this is equivalent but more robust than just explicitly defining 250 iterations)
num_iter = len(soup.select("div.chart-content.col-xs-12.col-sm-8 > p > em"))

tClist = soup.select("div.chart-content.col-xs-12.col-sm-8 > p > em")
spanlist = soup.select("div.chart-content.col-xs-12.col-sm-8 > p > cite")
# iterate through the result set and retrive all the data
for i in range(num_iter):
    artist.append(tClist[i].get_text())
    song.append(spanlist[i].get_text())

print(num_iter)
print(artist)
print(song)



# each list becomes a column
songs = pd.DataFrame({"artist":artist,
                       "song":song, 
                      })

songs

songs.head()

songs.dtypes

len(songs)

secure_random = random.SystemRandom()
print(secure_random.choice(songs['song']))

# print(songs['song'].sample())

favorite_song = input('Please enter your song name: ')


print(favorite_song)

# if favorite_song in songs['song']:
if len(songs[(songs['song'].isin([favorite_song]))])>0:
    print("Our song recommendation is: ", secure_random.choice(songs['song'])) 
else:
    print("We are sorry that we do not have a song recommendation for you now.")

songs_2=songs

## Top-100-Songs 2016

# 2. find url and store it in a variable
# top 100 songs - 2015
url = "https://playback.fm/charts/top-100-songs/2016"

# 3. download html with a get request
response = requests.get(url)
# response.status_code # 200 status code means OK!

response.status_code # 200 status code means OK!

# response.content

# 4.1. parse html (create the 'soup')
soup = BeautifulSoup(response.content, "html.parser")

# 4.2. check that the html code looks like it should
# soup

# 5. retrieve/extract the desired info (here, you'll paste the "Selector" you copied before to get the element that belongs to the top movie)
# soup.select("td:nth-child(2)")



# print(soup.prettify())

soup.select("td:nth-child(2) > a")[0]

soup.select("td:nth-child(2) > a")[99].get_text() # artist

soup.select("td:nth-child(2)")[0]

soup.select("td.mobile-hide > a > span.song")[0].get_text() #song

#initialize empty lists
artist = []
song = []

# define the number of iterations of our for loop
# by checking how many elements are in the retrieved result set
# (this is equivalent but more robust than just explicitly defining 250 iterations)
num_iter = len(soup.select("td:nth-child(2) > a"))

tClist = soup.select("td:nth-child(2) > a")
spanlist = soup.select("td.mobile-hide > a > span.song")
# iterate through the result set and retrive all the data
for i in range(num_iter):
    artist.append(tClist[i].get_text())
    song.append(spanlist[i].get_text())

print(num_iter)
print(artist)
print(song)



# each list becomes a column
songs = pd.DataFrame({"artist":artist,
                       "song":song, 
                      })

songs

songs['artist']=songs['artist'].str.replace('\n', ' ')
             

songs.head()

len(songs)

import random


secure_random = random.SystemRandom()
print(secure_random.choice(song))

# print(songs['song'].sample())

favorite_song = input('Please enter your song name: ')


print(favorite_song)

# Recommend song

if len(songs[(songs['song'].isin([favorite_song]))])>0:
    print("Our song recommendation is: ", secure_random.choice(songs['song'])) 
else:
    print("We are sorry that we do not have a song recommendation for you now.")

songs_3=songs

# CONCAT Songs_1, Song_2, Songs_3

songs_df = pd.concat([songs_1, songs_2, songs_3], ignore_index = True)
songs_df

songs_df.to_csv('songs_df.csv')

songs_df = pd.read_csv('songs_df.csv')

import random


secure_random = random.SystemRandom()
print(secure_random.choice(songs_df['song']))

favorite_song = input('Please enter your song name: ')


print(favorite_song)

# if favorite_song in songs['song']:
if len(songs_df[(songs_df['song'].isin([favorite_song]))])>0:
    print("Our song recommendation is: ", secure_random.choice(songs_df['song'])) 
else:
    print("We are sorry that we do not have a song recommendation for you now.")
